requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
openai
streamlit
pytest
